pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "doubaoWord2Anki"
version = "0.1.0"
description = "Import JSON data to ANKI with enhanced dictionary support"
authors = [{name = "pony"}]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "genanki>=0.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=0.21.0",
    "loguru>=0.6.0",
    "aiohttp>=3.8.0",
    "backoff>=2.2.0",
    "aiofiles>=23.1.0",
    "pyyaml>=6.0.0",
    "pytest>=8.3.5",
    "aioresponses>=0.7.8",
]

[tool.setuptools]
packages = ["src"]
package-dir = {"src" = "src"}

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "aioresponses>=0.7.4",
    "pytest-mock>=3.10.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[project.scripts]
anki-importer = "doubaoWord2Anki.cli:main"
